Cloud Router Extraction
https://harperdb.atlassian.net/browse/STUDIO-183
