+++ src/features/organization/OrganizationLayout.tsx
@@ -3,5 +3,3 @@ import { Outlet } from '@tanstack/react-router';
 export function OrganizationLayout() {
 	return <Outlet />;
 }
-
-
