+++ src/features/organizations/OrganizationsLayout.tsx
@@ -3,5 +3,3 @@ import { Outlet } from '@tanstack/react-router';
 export function OrganizationsLayout() {
 	return <Outlet />;
 }
-
-
