+++ src/features/profile/routes.ts
@@ -8,3 +8,13 @@
 // ];
 
 // export default [...profileRoutes];
+import { createRoute, lazyRouteComponent } from '@tanstack/react-router';
+import { dashboardLayout } from '@/router/dashboard-route';
+
+const profileRoute = createRoute({
+	getParentRoute: () => dashboardLayout,
+	path: 'profile',
+	component: lazyRouteComponent(async () => import('@/features/profile/index'), 'ProfileIndex'),
+});
+
+export const profileRouteTree = [profileRoute];
